+++ test/libsolidity/SMTChecker.cpp
@@ -35,12 +35,6 @@ namespace test
 
 class SMTCheckerFramework: public AnalysisFramework
 {
-public:
-	SMTCheckerFramework()
-	{
-		m_warningsToFilter.push_back("Experimental features are turned on.");
-	}
-
 protected:
 	virtual std::pair<SourceUnit const*, ErrorList>
 	parseAnalyseAndReturnError(
@@ -103,6 +97,7 @@ BOOST_AUTO_TEST_CASE(warn_on_struct)
 		}
 	)";
 	CHECK_WARNING_ALLOW_MULTI(text, (vector<string>{
+		"Experimental feature",
 		"Assertion checker does not yet implement this expression.",
 		"Assertion checker does not yet support the type of this variable."
 	}));